Add automatic update feed registration to deployment VSIX
This makes it more straightforward to stay up to date after an
initial install of a dogfooding VSIX. This works in conjuction
with the new VSIX feeds per feature branch, and removes the
need to manually configure a private VS gallery feed.

This works even when the VSIX is built locally, since a static
pkgdef can be referenced in the VSIX manifest even if it's not
present in the VSIX, in which case it will just be ignored.

The MoreInfo and ReleaseNotes URLs are provided to point to
the source commit and a comparison log for the branch too.

Also, given we stopped generating the InteractiveComponents VSIX
since #43208, we should not try to include it in the deployment
VSIX either, since this then fails to install because of the
missing payload.